Fix a bug in covlmc.prune.
The bug was triggered by situations when models have been simplified after their merging was rejected. During the post pruning phase, the new merging test was done assuming the models were not simplified, leading to a negative degree of freedom in the chi squared distribution.

The bug is solved by caching discarded models, a step in the direction of implementing improvement #15.