Add a new counting mode for contexts.*
In (co)vlmc models, a content may be valid and still have "descendants" in the context tree. The default reporting mode in contexts.* for those contexts is to count all occurences in the time series, even if some of them are matched by longer contexts. With counts="local", this default behaviour is replaced yb one in which only proper matches are attached to a given context.